requirements-unlocked.txt
boto3
celery
Flask
Flask-AlchemyDumps
Flask-Cors
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
gunicorn
psutil
psycopg2-binary
python-dotenv
recommonmark
redis
Sphinx
sphinx-rtd-theme
sphinxcontrib-httpdomain
sphinxcontrib-websupport
xlrd
requests

pycodestyle==2.3.1
pydocstyle==2.0.0
pylint==1.7.2
